Fix for LCM offsets shifting layout down
There's currently a bug where icons shift downwards across a row when
the Y coordinate that the icon starts on is not a multiple of its scale
factor. This doesn't happen when no icons are scaled, though it happens
frequently in sprite sheets with retina assets.

At 2X, an icon is only a multiple of its Y coordinate if the Y
coordinate is even. If the position of the icon is odd, it will shift
the icon down to the appropriate position. This incorrectly updated the
Y coordinate for the successive icons in the row, though, causing the
problem to cascade throughout the sheet. While the resulting behavior of
the sprite sheet is indeed correct (no icons are positioned incorrectly
on the rendered page), this does cause file sizes to grow unnecessarily.

While this doesn't account for a huge amount of waste, fixing it can
shave a decent number of kilobytes off of very large sprite sheets.